+++ CHANGELOG.md
@@ -9,6 +9,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 * Update library dependencies
 
+### Added
+
+* New process type `notify`
+  ([#41](https://gitlab.com/veenj/cinit/issues/41))
+
 ## [1.3.9]
 
 ### Added